Common component created
