+++ src/components/terminal/TerminalWrapper.tsx
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import { Terminal } from 'xterm';
-import { Scroller } from '../app/details/cIDetails/CIDetails'
 import { get } from '../../services/api';
 import { AppDetails } from '../app/types';
 import SockJS from 'sockjs-client';
@@ -13,6 +12,7 @@ import { SocketConnectionType } from '../app/details/appDetails/AppDetails';
 import { useThrottledEffect } from '../common';
 import ReactGA from 'react-ga4';
 import './terminal.css';
+import { Scroller } from '../app/details/cicdHistory/History.components';
 
 interface TerminalViewProps {
     appDetails: AppDetails;
